requirements.txt
torch
torchvision
timm
flask
flask-cors
albumentations
pillow
numpy
gunicorn